Updated publisher - use topic.
