+++ EVCP.DataConsumer/Publisher/PublishWorker.cs
@@ -4,16 +4,16 @@ namespace EVCP.DataConsumer.Publisher;
 
 public class PublishWorker : IWorker
 {
-    private readonly IEVDataPublisher _publisher;
+    private readonly ITripDataPublisher _publisher;
     private readonly string _routingKey;
 
-    private readonly Func<IEnumerable<ITripDataDto>> _generateMessages;
+    private readonly Func<IEnumerable<ITripDataDto>> _getMessages;
 
-    public PublishWorker(IEVDataPublisher publisher, string routingKey, Func<IEnumerable<ITripDataDto>> generateMessages)
+    public PublishWorker(ITripDataPublisher publisher, string routingKey, Func<IEnumerable<ITripDataDto>> getMessages)
     {
         _publisher = publisher ?? throw new ArgumentNullException(nameof(publisher));
         _routingKey = routingKey;
-        _generateMessages = generateMessages;
+        _getMessages = getMessages;
     }
 
     public async Task Run()
@@ -23,7 +23,7 @@ public class PublishWorker : IWorker
 
     private async Task ProcessFile()
     {
-        var messages = _generateMessages();
+        var messages = _getMessages();
 
         messages.ToList().ForEach(async data => await _publisher.Publish(data, _routingKey));
     }
